Add: handling of grouped data and attributes
If data is related, like first and last name, they should always be
printed together. But if one is missing the two will become
misaligned. Grouping data allows us to tell the computer they are
connected and print them both in the same row of a file. If one is
missing they will be printed as an empty string.