Add: yaml reader to move node data to yaml file
Allows node data to be stored in a hierarchy in a structure file that
is then used to create the node structures instead of having the
values hard coded.

This allows:
- Separation of data + code.
- Changing node values without having to recompile.
- And allows other programs to read the same structure
data. (i.e. will allow a shell script to also read the structure file
and use that to create the final node files from the cache files.)