Add: two more tests
It does not appear the method for sorting makes a difference. Each
finishes in essentially the same amount of time.

Should still be tested using >> 4 threads to see if there is a greater
difference. Don't know why the method of concatenating all files
together first should be as fast as the other methods since it seems
like it should not be able to spread the load over processors like the
multi-file methods can.