Optimize: reduce comparisons by using symmetry of the overlap matrix
Add in a check to ensure the primary column is sorted. (Needed for
using symmetry but also needed so that the last line of the contains
the largest ID anyway.) If the primary column is sorted, only need to
check lines below it since overlap matrix is
symmetric (`overlap[smaller][larger] == overlap[larger][smaller]`) and
only the first case is printed.

Might be useful to keep track of the diagonal values for normalizing
but that can be added in later if needed easily.

Also did some cleaning by reorder the code removing the need to rewind
and skip lines in a few places.