Remove: printing headers
The neo4j syntax for headers makes this difficult since the first two
columns of the output are of the primary columns node type so we
should get the first columns header and print in both places but
with neo4j syntax that leads to two `:START_ID(%s)` instead of
`:START_ID(%s)\t:END_ID(%s)`. Instead of coming up with an over
complicated method of doing this it's easier to just generate the
headers in a bash script.