Optimize: reduce memory requirements
Only keep track of a single node's overlap at a time. Since the
primary column most be sorted and overlap is symmetric, once the
primary column has moved past the current node the current node's
overlap is finalized so it can be printed and forgotten. This allows
for only storing a 1d array of size n_nodes instead of a n_nodes x
n_nodes 2d array.

Additionally, moved sort check outside of main loop. This simplifies
the main loop, reduces repeated checks, and causes a quick error if
the primary column is not sorted.