Change: logic for determining if in header
`atoi` returns 0 if it can't parse the string to an integer but
0 is a possible node id value making it difficult to determine if a
zero corresponds to an actual 0 or a non-integer value. This works
fine when assuming nodes will always have positive ids. Switching to
break out f `skip_header` if one of the columns is non-zero allows for
handling zero valued node ids but not when both columns are zero.