Optimize: rewrite to allow for vector optimizations
Now loop over pairs of nodes instead of pairs of edges. So run over
all pairs of edges for the given node pair. This allows for `overlap`
to be a scalar that is calculated and printed for each node
pair. However, I changed it back to an array since reducing the print
calls (in the `OMP CRITICAL` block) improved
performance.

Additionally, this should reduce number of iterations
slightly since it's now going over the true number of nodes rather
than the expected number of nodes which assumes nodes are perfectly
sequentially. There have been a small number, ~1%, of node ids that
are missing but I'm not sure why yet.