Optimize: remove omp critical pragma
Thought this was needed to prevent interleaving when multiple
processes write to the same stream. But `printf` handles this itself
as long as the amount to be written to the stream is smaller than the
streams buffer size. Removing this command significantly reduces
runtime since threads don't have to wait for `flush_overlap` to be
free.