Fix: sort overlap test output
Using parallel processing leads to disordered output. This is fine in
general since neo4j's importer is also parallel and will not read the
edges in order anyway. But to get the tests to pass they most be in
the expected order so sort the outputs to ensure the calculated
weights are correct.