Fix: incorrect concatenation of files
Each processor writes to it's own file then the files are concatenated
into a single file at the end. Because the processor files were not
being closed at the end they did not do their final write out so the
last bit of data is not written until the process finishes, which was
after concatenation.

As part of the figuring out the problem I went ahead and wrote cat in
C to replace the system call so no longer dependent on coreutils
being on the path.