Fix error with repeated calls to `parse_file`
Was reusing the same memory for the tag buffer which meant the value was preset
to the value of the last tag in the previous call which would be the root
tag. This caused the current tag to match root before it had really found the
root tag. Make sure to preallocate strings.